refactor: extract utility functions to improve code organization
- Move createBatches and retryOperation to new utils.ts file
- Remove separate batches.ts and retry.ts files
- Improve code readability by reducing nested functions in pipeline.ts
- Also includes minor refactoring in route.ts from earlier changes
